refactor(earn/poolcardtri directory): delete unused comps. Move PoolCardTri compsto own directory
