+++ src/components/earn/PoolCardTri/PoolCardTriRewardText.tsx
@@ -1,16 +1,16 @@
 import React, { useCallback, useMemo, useState } from 'react'
 import { ChainId, Token } from '@trisolaris/sdk'
-import { TYPE } from '../../theme'
-import { AutoColumn } from '../Column'
-import { AutoRow, RowBetween } from '../Row'
-import { TRI } from '../../constants/tokens'
-import Popover from '../Popover'
-import { PoolCardTriProps } from './PoolCardTri'
+import { TYPE } from '../../../theme'
+import { AutoColumn } from '../../Column'
+import { AutoRow, RowBetween } from '../../Row'
+import { TRI } from '../../../constants/tokens'
+import Popover from '../../Popover'
+import { PoolCardTriProps } from '../PoolCardTri'
 import styled from 'styled-components'
 import { Info } from 'react-feather'
 import _ from 'lodash'
-import CurrencyLogo from '../CurrencyLogo'
-import useGetTokenByAddress from '../../hooks/useGetTokenByAddress'
+import CurrencyLogo from '../../CurrencyLogo'
+import useGetTokenByAddress from '../../../hooks/useGetTokenByAddress'
 
 const IconWrapper = styled.div`
   ${({ theme }) => theme.flexColumnNoWrap};
