+++ src/pages/EarnTri/ManageImpl.tsx
@@ -6,9 +6,9 @@ import { AutoColumn } from '../../components/Column'
 import { AutoRow, RowBetween } from '../../components/Row'
 import { CardSection, HighlightCard } from '../../components/earn/styled'
 import { ButtonPrimary } from '../../components/Button'
-import StakingModal from '../../components/earn/StakingModalTri'
-import UnstakingModal from '../../components/earn/UnstakingModalTri'
-import ClaimRewardModal from '../../components/earn/ClaimRewardModalTri'
+import StakingModal from '../../components/earn/PoolCardTri/StakingModalTri'
+import UnstakingModal from '../../components/earn/PoolCardTri/UnstakingModalTri'
+import ClaimRewardModal from '../../components/earn/PoolCardTri/ClaimRewardModalTri'
 import { PageWrapper } from '../../components/Page'
 import CountUp from '../../components/CountUp'
 import MultipleCurrencyLogo from '../../components/MultipleCurrencyLogo'
