+++ src/pages/StakeTri/AboutContainer.tsx
@@ -2,7 +2,7 @@ import React from 'react'
 import styled from 'styled-components'
 import { LightCard } from '../../components/Card'
 import { AutoColumn } from '../../components/Column'
-import PoolCardTriRewardText from '../../components/earn/PoolCardTriRewardText'
+import PoolCardTriRewardText from '../../components/earn/PoolCardTri/PoolCardTriRewardText'
 import { StableSwapPoolName, STABLESWAP_POOLS } from '../../state/stableswap/constants'
 import { useFarmsAPI } from '../../state/stake/useFarmsAPI'
 import { ExternalLink, TYPE } from '../../theme'
